Update `requirements.txt`
I removed 'exceptiongroup', which isn't needed in Python 3.11 or later;
and 'waitress', which was unused. I added 'PyQt5-Qt5' and 'PyQt5-sip',
which are dependencies of 'PyQt5'; 'setuptools', which is a dependency
of 'pyinstaller' and isn't preinstalled by `venv` since Python 3.12; and
'websocket-client', which is a dependency of 'selenium' since 4.22.0.

I would like to add version numbers and hashes to `requirements.txt`,
but I haven't done so yet since I don't want to make too many changes.