Replace `undetected-chromedriver`
We're encountering bugs in `undetected-chromedriver`, and the library
hasn't been significantly updated in over a year. Due to the lack of
maintenance, I think we should use a different library: SeleniumBase.

SeleniumBase includes a modified version of `undetected-chromedriver`,
making it a near drop-in replacement except for a different constructor.

SeleniumBase can also automatically download ChromeDriver. We set
`driver_version` to "mlatest" to download the latest ChromeDriver for
the current major version, just like `chromedriver-autoinstaller`.

The remaining changes are minor: we must explicitly set the window size
to match `undetected-chromedriver`, and we must look for "uc_driver.exe"
instead of "undetected_chromedriver.exe" when killing processes.