Remove unnecessary requirements
`eventlet`, `psutil`, and `PySocks` are all optional. We don't use the
`eventlet` async mode in `app.py`, so we don't need `eventlet`; we don't
monitor `gevent`, so we don't need `psutil`; and `PySocks` is already
required by a dependency, so we don't need to directly require it.

P.S.: If we want to monitor `gevent`, change the `gevent` requirement to
`gevent[monitor]`. The brackets denote "extras": optional components of
a dependency. Using an extra is better than directly requiring `psutil`.