Fix attribute formatting in `CLEAN_format_data_to_excel`
`zip` iterates over arrays in parallel, returning a tuple for each set
of elements at the same index. The issue in our code was that wherever
`zip` was used to iterate over a single array, the whole tuple was
converted to a string instead of the element inside the tuple, producing
cells that look like `('555-5555',)` instead of `555-5555`.

One solution is to add a trailing comma per the Python documentation:
<https://docs.python.org/3/reference/simple_stmts.html#assignment-statements>

The simpler solution is to remove the unnecessary calls to `zip`.