Don't join the selected keywords into comma-separated strings
This fixes a bug where an empty selection of keywords was joined into an
empty string, which was split by the server into an array containing an
empty string, *not* an empty array. Although possibly counterintuitive,
the empty string edge case in `str.split` is well-documented:

<https://docs.python.org/3/library/stdtypes.html#str.split>

This used to be a minor issue, but now the empty "keyword" is violating
CHECK constraints in the database, stopping the scraper prematurely.