Increase the maximum length of `varchar` columns
We already increased the maximum length of `description` in
`raw_mega_personals_posts`, so to avoid truncation errors, we'll go
ahead and increase the maximum length of every column.

We could truncate the data before sending it to the database, but we'll
need a way for the server to discover the maximum length of each column.