Store provider IDs in RubRatings as (non-null) strings
Every post has a provider ID, but the previous locator was buggy, so we
had to allow null values into the `provider_id` column. Now that we've
fixed the locator, we can add `NOT NULL`, and since we're already here,
we'll change the column to use a more appropriate data type, too.