Refactor results folder selection
This commit changes the client to use Electron's `dialog` module instead
of relying on the server to display a file dialog. I also deleted some
repetitive code by removing the 'result_folder_selected' event.

The `QFileDialog` class was the only feature we needed from Qt, so we
can now remove NetSpider's dependency on Qt 5 and PyQt5.